feat(helper): export service extensions mixin
