+++ packages/fdb_helper/lib/fdb_helper.dart
@@ -1,7 +1,8 @@
 /// Helper package for fdb (Flutter Debug Bridge).
 ///
-/// Add [FdbBinding.ensureInitialized] as the first line of your `main()` to
-/// register VM service extensions for widget interaction.
+/// Use [FdbBinding.ensureInitialized] or add [FdbServiceExtensionsMixin] to a
+/// custom binding to register VM service extensions for widget interaction.
 library;
 
 export 'src/fdb_binding.dart';
+export 'src/fdb_service_extensions_mixin.dart';
