+++ packages/fdb_helper/pubspec.yaml
@@ -16,6 +16,8 @@ dependencies:
   shared_preferences: ^2.0.0
 
 dev_dependencies:
+  flutter_test:
+    sdk: flutter
   flutter_lints: ^6.0.0
   pigeon: ^26.3.4
